split into spring and easing controls
